Code/Device/requirements.txt
# Pothole Detection System - Raspberry Pi Requirements


ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0

RPi.GPIO>=0.7.1
pyserial>=3.5

requests>=2.31.0
urllib3>=2.0.0

pandas>=2.0.0
matplotlib>=3.7.0

colorlog>=6.7.0
python-dotenv>=1.0.0

gpsd-py3>=0.3.0

pygame>=2.5.0
pyaudio>=0.2.11

pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
